Move SearchData into it's own file.
