+++ src/api/offices.ts
@@ -1,5 +1,6 @@
+import { SearchData } from 'src/types/SearchData'
 import apiClient from './client'
-import { SearchData } from './search'
+import Office from 'src/types/Office'
 
 export function get(id: string) {
   return apiClient
@@ -13,6 +14,8 @@ interface IOfficeList {
   page?: number
 }
 
-export function list(params: IOfficeList = {}): Promise<SearchData> {
-  return apiClient.get('offices', { params }).then((r) => r.data as SearchData)
+export function list(params: IOfficeList = {}): Promise<SearchData<Office>> {
+  return apiClient
+    .get('offices', { params })
+    .then((r) => r.data as SearchData<Office>)
 }
