+++ src/types/SearchData.ts
@@ -1,0 +1,6 @@
+import SearchMeta from './SearchMeta'
+
+export type SearchData<T> = {
+  meta: SearchMeta
+  data: T[]
+}
